[OTA-703] Reorder wizard steps
