+++ ota-plus-web/app/reactapp/src/components/campaigns/wizardSteps/index.js
@@ -1,6 +1,6 @@
 import WizardStep1 from './Step1';
-import WizardStep2 from './Step2';
 import WizardStep3 from './Step3';
+import WizardStep2 from './Step2';
 import WizardStep4 from './Step4';
 import WizardStep5 from './Step5';
 import WizardStep6 from './Step6';
