+++ ota-plus-web/app/reactapp/src/components/campaigns/wizardSteps/step2Files/GroupsList.jsx
