+++ ota-plus-web/app/reactapp/src/components/campaigns/wizardSteps/step3Files/index.js
@@ -1,7 +1,7 @@
-import WizardGroupsList from './GroupsList';
-import WizardGroupsListItem from './GroupsListItem';
+import SelectUpdateList from './SelectUpdateList';
+import UpdateDetails from '../../../updates/UpdateDetails';
 
 export {
-    WizardGroupsList,
-    WizardGroupsListItem
-}
+    SelectUpdateList,
+    UpdateDetails,
+};
